\name{MW}
\alias{MW}
\title{Get Frequency Data
}
\description{Orthogonal matrix defined in Harvey (1978)
}
\usage{
MW(n)
}
\arguments{
  \item{n}{rows and columuns number  
}
}
\value{Orthogonal matrix of n X n dimensions
}
\references{Harvey, A.C. (1978), Linear Regression in the Frequency Domain, International Economic Review, 19, 507-512.

%% ~put references to the literature/web site here ~
}
\author{Francisco Parra 
%%  ~~who you are~~
}
\examples{
MW(80)
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
}
\keyword{smooth}
\seealso{\code{\link[descomponer]{gdt},\link[descomponer]{gdf},\link[descomponer]{cdf}}}